+++ src/interfaces/ClassType.ts
@@ -1,0 +1,3 @@
+export interface ClassType<T = any> {
+  new (): T;
+}