+++ src/interfaces/index.ts
@@ -1,5 +1,6 @@
 export { MiddlewareFn, NextFn, MiddlewareInterface } from "./Middleware";
 export * from "./AuthChecker";
+export * from "./ClassType";
 export * from "./Publisher";
 export * from "./ResolverData";
 export * from "./ResolverFilterData";